requirements.txt
aiohttp
kafka-python
redis
pandas
numpy
python-dotenv
loguru
python-json-logger
fake-useragent
yfinance
praw
beautifulsoup4

vaderSentiment
textblob
pyarrow
fastparquet
scikit-learn
mlflow
optuna
xgboost
torch
torchvision
torchaudio

prometheus_client
fastapi
uvicorn
python-multipart
httpx
pydantic